Combine utils tests as one and add tests for combine_expressions #1492
